requirements.txt
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
scikit-learn==1.3.2
